Start using the EFI stub and UEFI boot
The Linux kernel can be directly booted from UEFI by using the EFI stub.
The challenge is getting the initramfs to be loaded. The EFI stub scans
the Linux commandline when started for `initrd=` and will call UEFI APIs
to load it. This happens very early and depends on UEFI passing the
parameters. This can be done with the UEFI shell (included and used via
startup.nsh) or by updating the NVRAM. Updating NVRAM is done via
efibootmgr.