pstore: enable persistent logging to DRAM
This adds support for the RamoopsLogger backend for preserving log
messages between unintended reboots. It reserves a 256KB region near the
beginning of DRAM (1 MB offset) for the pmsg circular buffer. The pstore
drivers were already enabled. This tightens the configuration options by
removing the EFI var pstore backend which can't be used to store pmsgs.

NOTE: KASLR (address space randomization) was turned off due to advice
from
https://docs.pmem.io/persistent-memory/getting-started-guide/creating-development-environments/linux-environments/linux-memmap.
The note at the bottom implies that Linux ignores the `memmap`
commandline parameter when it comes to KASLR. I did not experiment
enough to know whether this is true or not, so I left it out.